requirements.txt
Flask==2.3.2
pandas==2.2.3
python-dotenv==1.0.0
geopy==2.4.0
requests==2.31.0
